Merged PR 19118: Replace <depends> tag for Shared Catalog sync toggle
## What's being changed

We're changing the way we show/hide the Shared Catalog Sync Enabled select in Sync Settings.

## Why it's being changed

By adding the `<depends>` tag for Shared Catalog Enabled, to make it dependent on regular Catalog Sync Enabled, we accidentally broke the 'disabled' flag that is added to Shared Catalog Enabled if B2B Features > Shared Catalogs is disabled.

This PR removes the `<depends>` tag and adds more logic to the front end model, to disable the field if either B2B Features > Enable Shared Catalog or Sync Settings > Catalog Enabled are not active.

## How to review / test this change

- Disable B2B Features > Shared Catalog
- Disable Sync Settings > Catalog Enabled
- Confirm that Sync Settings > Shared Catalog Enabled is disabled
- Check the comment indicates that two other things need to be enabled before this field is accessible
- Enable the blockers one by one and observe the comment text changing

Related work items: #107379