Merged PR 49720: Handle company status not matching expected values
## What's being changed

Our CustomerPlugin.

## Why it's being changed

Our code was looking for a match between the company status of the contact being synced, and [a limited set of statuses](https://github.com/dotmailer/dotmailer-magento2-extension-b2b/blob/master/Helper/ConfigInterface.php#L64-L69). These are based on values defined in `Magento\Company\Api\Data\CompanyInterface`, but evidently there can be variations. This PR will sync either a matching status, if we can find one, or if not, the status as it comes to us.

## How to review / test this change

- Enable B2B in your Magento
- Create a company
- Create a customer
- Assign the customer to the company
- Map the 'company status' data field
- Run customer sync
- Check the export, and see a company status synced as expected
- Hack the code and test if `$company->getStatus()` returns null, or is a value not included in our array of statuses

Related work items: #234247